pyproject.toml
[tool.poetry]
name = "django-multi-pos"
version = "0.1.0"
description = ""
authors = ["William De la Cruz <macwilliamdlc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
uritemplate = "^3.0.1"
pyyaml = "^5.3.1"
factory_boy = "^3.0.1"
drf-yasg = "^1.17.1"
pydantic = "^1.8.1"
pytest = "^6.2.2"
pytest-django = "^4.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
